% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{add_partition_split}
\alias{add_partition_split}
\title{Adds partition_split to grid_partition}
\usage{
add_partition_split(obj, s)
}
\arguments{
\item{obj}{Grid Partition object}

\item{s}{Partition Split object}
}
\value{
updated Grid Partition
}
\description{
Adds partition_split to grid_partition
}
