% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalGrid.R
\docType{package}
\name{CausalGrid-package}
\alias{CausalGrid}
\alias{CausalGrid-package}
\title{CausalGrid: Analysis of Subgroups}
\description{
To learn more about dplyr, start with the vignettes:
`browseVignettes(package = "CausalGrid")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://microsoft.github.io/CausalGrid/}
  \item \url{https://github.com/microsoft/CausalGrid}
}

}
\author{
\strong{Maintainer}: Brian Quistorff \email{Brian.Quistorff@microsoft.com}

}
\keyword{internal}
