% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalGrid.R
\docType{package}
\name{CausalGrid}
\alias{CausalGrid}
\title{CausalGrid: A package for subgroup effects}
\description{
Intervals are (a,b], and [a,b] for the lowest. A split at x means <= and >
We randomize in generating train/est and trtr/trcv splits. Possibly cv.glmnet and cv.gglasso as well.
}
