% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{change_complexity}
\alias{change_complexity}
\title{Change the complexity of a fit_estimate_partition}
\usage{
change_complexity(fit, y, X, d = NULL, partition_i)
}
\arguments{
\item{fit}{estimated_partition}

\item{y}{Nx1 matrix of outcome (label/target) data. With multiple core estimates see Details below.}

\item{X}{NxK matrix of features (covariates). With multiple core estimates see Details below.}

\item{d}{(Optional) NxP matrix (with colnames) of treatment data. If all equally important they 
should be normalized to have the same variance. With multiple core estimates see Details below.}

\item{partition_i}{partition_i - 1 is the last include in split_seq included in new partition}
}
\value{
updated estimated_partition
}
\description{
Change the complexity level of the partition and re-estimate cell statistics.
}
\details{
Note: doesn't update the importance weights
}
