% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{const_vect}
\alias{const_vect}
\title{Determine if vector is constant (or zero length)}
\usage{
const_vect(var)
}
\arguments{
\item{var}{A numeric vector}
}
\description{
Determine if vector is constant (or zero length)
}
\keyword{internal}
