% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{const_vectr}
\alias{const_vectr}
\title{inline', second just Rcpp), but couldn't get to work in building a package.}
\usage{
const_vectr(x)
}
\description{
inline', second just Rcpp), but couldn't get to work in building a package.
}
