% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Do_Residualize.grid_rf}
\alias{Do_Residualize.grid_rf}
\title{Do_Residualize.grid_rf}
\usage{
\method{Do_Residualize}{grid_rf}(obj, y, X, d, sample)
}
\arguments{
\item{obj}{Object}

\item{y}{y}

\item{X}{X}

\item{d}{d (Default=NULL)}

\item{sample}{one of 'tr' or 'est'}
}
\value{
list(y=) or list(y=, d=)
}
\description{
Do_Residualize.grid_rf
}
