% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Do_Residualize.lm_X_est}
\alias{Do_Residualize.lm_X_est}
\title{Do_Residualize.lm_X_est}
\usage{
\method{Do_Residualize}{lm_X_est}(obj, y, X, d, sample)
}
\arguments{
\item{obj}{obj}

\item{y}{y}

\item{X}{X}

\item{d}{d}

\item{sample}{one of 'tr' or 'est'}
}
\value{
list(y=...) or list(y=..., d=...)
}
\description{
Do_Residualize.lm_X_est
}
