% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{est_params}
\alias{est_params}
\alias{est_params.lm_est}
\alias{est_params.simple_est}
\alias{est_params.grid_rf}
\title{Estimate parameters}
\usage{
est_params(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)

\method{est_params}{lm_est}(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)

\method{est_params}{simple_est}(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)

\method{est_params}{grid_rf}(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)
}
\arguments{
\item{obj}{an EstimatorPlan object}

\item{y}{A N-vector}

\item{d}{A N-vector or NxM matrix (so that they can be estimated jointly)}

\item{X}{A NxK matrix or data.frame}

\item{sample}{One of: "trtr", "trcv", "est"}

\item{ret_var}{Return Variance in the return list}
}
\value{
\code{list(param_est=...)} or \code{list(param_est=...)} if \code{ret_var}
}
\description{
Estimate parameters on the data.
}
\section{Methods (by class)}{
\itemize{
\item \code{lm_est}: lm_est

\item \code{simple_est}: simple_est

\item \code{grid_rf}: grid_rf
}}

