% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_plans.R
\name{EstimatorPlan}
\alias{EstimatorPlan}
\title{Estimator Plan}
\description{
An estimator plan is an S3 objet that defines functions for how to estimate parameters in the various stages of subgroup analysis.
}
\details{
With an outcome-only model, then we just call \code{\link{est_params}}. 
When there is a treatment, then we initially also call \code{\link{fit_on_train}}, \code{\link{residualize}}.

Each plan also has a field \code{dof} for a degrees of freedom-correction to standard errors (e.g. 2 (intercept and treatment effect) or more if controls are included) 
Provided estimator plans include \code{\link{simple_est}}, \code{\link{lm_est}}, and \code{\link{grid_rf}}. User provided plans just need to implement the relevant functions.
}
