% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{fit_estimate_partition}
\alias{fit_estimate_partition}
\title{fit_estimate_partition}
\usage{
fit_estimate_partition(
  y,
  X,
  d = NULL,
  max_splits = Inf,
  max_cells = Inf,
  min_size = 3,
  cv_folds = 2,
  potential_lambdas = NULL,
  lambda.1se = FALSE,
  partition_i = NA,
  tr_split = 0.5,
  verbosity = 0,
  pot_break_points = NULL,
  bucket_min_n = NA,
  bucket_min_d_var = FALSE,
  honest = FALSE,
  ctrl_method = "",
  pr_cl = NULL,
  alpha = 0.05,
  bump_B = 0,
  bump_ratio = 1,
  importance_type = ""
)
}
\arguments{
\item{y}{N vector of outcome (label/target) data}

\item{X}{NxK matrix of features (covariates). Must be numerical (unordered categorical variables must be 
1-hot encoded.)}

\item{d}{(Optional) N vector of treatment data.}

\item{max_splits}{Maximum number of splits even if splits continue to improve OOS fit}

\item{max_cells}{Maximum number of cells}

\item{min_size}{Minimum size of cells}

\item{cv_folds}{Number of CV Folds or foldids. If Multiple effect #3 and using vector, then pass in list of vectors.}

\item{potential_lambdas}{potential lambdas to search through in CV}

\item{lambda.1se}{Use the 1se rule to pick the best lambda}

\item{partition_i}{Default is NA. Use this to avoid CV automated selection of the partition}

\item{tr_split}{- can be ratio or vector of indexes. If Multiple effect #3 and using vector then pass in list of vectors.}

\item{verbosity}{If >0 prints out progress bar for each split}

\item{pot_break_points}{NULL or a k-dim list of vectors giving potential split points for non-categorical 
variables (can put c(0) for categorical). Similar to 'discrete splitting' in 
CausalTree though their they do separate split-points for treated and controls.}

\item{bucket_min_n}{Minimum number of observations needed between different split checks for continuous features}

\item{bucket_min_d_var}{Ensure positive variance of d for the observations between different split checks 
for continuous features}

\item{honest}{Whether to use the emse_hat or mse_hat. Use emse for outcome mean. For treatment effect, 
use if want predictive accuracy, but if only want to identify true dimensions of heterogeneity 
then don't use.}

\item{ctrl_method}{Method for determining additional control variables. Empty ("") for nothing, "all" or "lasso"}

\item{pr_cl}{Parallel Cluster (If NULL, default, then will be single-processor)}

\item{alpha}{Default=0.05}

\item{bump_B}{Number of bump bootstraps}

\item{bump_ratio}{For bootstraps the ratio of sample size to sample (between 0 and 1, default 1)}

\item{importance_type}{Options:
single - (smart) redo full fitting removing each possible dimension
interaction - (smart) redo full fitting removing each pair of dimensions
 "" - Nothing}
}
\value{
An object with class \code{"estimated_partition"}.
\item{partition}{Parition obj defining cuts}
\item{cell_stats}{list(cell_factor=cell_factor, stats=stat_df) from estimate_cell_stats() using est sample}
\item{importance_weights}{importance_weights}
\item{interaction_weights}{interaction_weights}
\item{has_d}{has_d}
\item{lambda}{lambda used}
\item{is_obj_val_seq}{In-sample objective function values for sequence of partitions}
\item{complexity_seq}{Complexity #s (# cells-1) for sequence of partitions}
\item{partition_i}{Index of Partition selected in sequence}
\item{split_seq}{Sequence of splits. Note that split i corresponds to partition i+1}
\item{index_tr}{Index of training sample (Size of N)}
\item{cv_foldid}{CV foldids for the training sample (Size of N_tr)}
\item{varnames}{varnames (or c("X1", "X2",...) if X doesn't have colnames)}
\item{honest}{honest target}
\item{est_plan}{Estimation plan}
\item{full_stat_df}{full_stat_df}
}
\description{
Split the data, one one side train/fit the partition and then on the other estimate subgroup effects.
With multiple treatment effects (M) there are 3 options (the first two have the same sample across treatment effects).
 1) Multiple pairs of (Y_{m},W_{m}). y,X,d are then lists of length M. Each element then has the typical size
    The N_m may differ across m. The number of columns of X will be the same across m.
 2) Multiple treatments and a single outcome. d is then a NxM matrix.
 3) A single treatment and multiple outcomes. y is then a NXM matrix.
}
