% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Fit_InitTr.grid_rf}
\alias{Fit_InitTr.grid_rf}
\title{Fit_InitTr.grid_rf
Note that for large data, the rf_y_fit and potentially rf_d_fit objects may be large.
They can be null'ed out after fitting}
\usage{
\method{Fit_InitTr}{grid_rf}(
  obj,
  X_tr,
  y_tr,
  d_tr = NULL,
  cv_folds,
  verbosity = 0,
  dim_cat = c()
)
}
\arguments{
\item{obj}{Object}

\item{X_tr}{X}

\item{y_tr}{y}

\item{d_tr}{d_tr}

\item{cv_folds}{CV folds}

\item{verbosity}{verbosity}

\item{dim_cat}{vector of dimensions that are categorical}
}
\value{
Updated Object
}
\description{
Fit_InitTr.grid_rf
Note that for large data, the rf_y_fit and potentially rf_d_fit objects may be large.
They can be null'ed out after fitting
}
