% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Fit_InitTr}
\alias{Fit_InitTr}
\title{Fit_InitTr}
\usage{
Fit_InitTr(
  obj,
  X_tr,
  y_tr,
  d_tr = NULL,
  cv_folds,
  verbosity = 0,
  dim_cat = c()
)
}
\arguments{
\item{obj}{Object}

\item{X_tr}{X}

\item{y_tr}{y}

\item{d_tr}{d_tr}

\item{cv_folds}{CV folds}

\item{verbosity}{verbosity}

\item{dim_cat}{vector of dimensions that are categorical}
}
\value{
Updated Object
}
\description{
Fit_InitTr
}
