% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Fit_InitTr.simple_est}
\alias{Fit_InitTr.simple_est}
\title{Fit_InitTr.simple_est}
\usage{
\method{Fit_InitTr}{simple_est}(
  obj,
  X_tr,
  y_tr,
  d_tr = NULL,
  cv_folds,
  verbosity = 0,
  dim_cat = c()
)
}
\arguments{
\item{obj}{obj}

\item{X_tr}{X_tr}

\item{y_tr}{y_tr}

\item{d_tr}{d_tr}

\item{cv_folds}{cv_folds}

\item{verbosity}{verbosity}

\item{dim_cat}{dim_cat}
}
\value{
Updated object
}
\description{
Fit_InitTr.simple_est
}
