% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{fit_on_train}
\alias{fit_on_train}
\title{Fit estimator_plan on training sample}
\usage{
fit_on_train(obj, X_tr, y_tr, d_tr, cv_folds, verbosity = 0, dim_cat = c())
}
\arguments{
\item{obj}{an \code{EstimatorPlan} object}

\item{X_tr}{N_trxK matrix}

\item{y_tr}{N_tr-vector}

\item{d_tr}{NULL, N_tr-vector, or N_trxM matrix (if multiple treatments)}

\item{cv_folds}{CV folds}

\item{verbosity}{verbosity}

\item{dim_cat}{vector of dimensions that are categorical}
}
\value{
Updated EstimatorPlan object
}
\description{
Initializes the estimator (if necessary) on the \code{train} sample.
}
