% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{get_desc_df.grid_partition}
\alias{get_desc_df.grid_partition}
\title{Get descriptive data.frame for grid_partition}
\usage{
get_desc_df.grid_partition(
  partition,
  cont_bounds_inf = TRUE,
  do_str = FALSE,
  drop_unsplit = FALSE,
  digits = NULL,
  unsplit_cat_star = TRUE
)
}
\arguments{
\item{partition}{Partition}

\item{cont_bounds_inf}{If True, will put continuous bounds as -Inf/Inf. Otherwise will use X_range bounds}

\item{do_str}{If True, use a string like "(a, b]", otherwise have two separate columns with a and b}

\item{drop_unsplit}{If True, drop columns for variables overwhich the partition did not split}

\item{digits}{digits option}

\item{unsplit_cat_star}{if we don't split on a categorical var, should we show as "*" (otherwise list all levels)}
}
\value{
data.frame
}
\description{
A dataset with rows for each cell and columns defining partitioning
}
