% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{get_factor_from_partition}
\alias{get_factor_from_partition}
\title{Get factor describing cell number fo each observation}
\usage{
get_factor_from_partition(partition, X, X_range = NULL)
}
\arguments{
\item{partition}{partition}

\item{X}{X data or list of X}

\item{X_range}{(Optional) overrides the partition$X_range}
}
\value{
Factor
}
\description{
Note that currently if X has values more extreme (e.g., for numeric or factor levels ) than was used to generate the partition
then we will return NA unless you provide and updated X_range.
}
