% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{get_X_range}
\alias{get_X_range}
\title{Get X_range}
\usage{
get_X_range(X)
}
\arguments{
\item{X}{data}
}
\value{
list of length K with each element being a c(min, max) along that dimension
}
\description{
Get X_range
}
