% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{grid_partition}
\alias{grid_partition}
\alias{is_grid_partition}
\title{Create a null \code{grid_partition}}
\usage{
grid_partition(X = NULL, X_range = NULL, varnames = NULL)

is_grid_partition(x)
}
\arguments{
\item{X}{Feature data. Must pass in this or \code{X_range}.}

\item{X_range}{Such as from \code{\link{get_X_range}} (saves time if already computed)}

\item{varnames}{Names of the X-variables}

\item{x}{an R object}
}
\value{
Grid Partition

True if x is a grid_partition
}
\description{
Create a empty partition. Splits can be added using \code{\link{add_partition_split}}.
Information about a split can be retrieved using \code{\link{num_cells}}, \code{\link{get_desc_df}} and \code{\link{print}}
With data, one can determine the cell for each observation using \code{\link{predict}}

Test whether an object is an \code{grid_function}
}
\section{Functions}{
\itemize{
\item \code{is_grid_partition}: is grid_partition
}}

