% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{is.grid_partition_split}
\alias{is.grid_partition_split}
\title{Is grid_partition_split}
\usage{
is.grid_partition_split(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Boolean
}
\description{
Is grid_partition_split
}
\section{Functions}{
\itemize{
\item \code{is.grid_partition_split}: is grid_partition_split
}}

