% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{GridPartition}
\alias{GridPartition}
\title{Grid Partition}
\description{
A \code{\link{grid_partition}} is defines a grid over a feature-space. It can be built by composing \code{\link{partition_split}}s.
}
\details{
The partition is typically built by a search algorithm as \code{\link{fit_partition}}.
}
