% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{is.estimated_partition}
\alias{is.estimated_partition}
\title{is.estimated_partition}
\usage{
is.estimated_partition(x)
}
\arguments{
\item{x}{Object}
}
\value{
True if x is an estimated_partition
}
\description{
is.estimated_partition
}
