% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{is.grid_partition}
\alias{is.grid_partition}
\title{is.grid_partition}
\usage{
is.grid_partition(x)
}
\arguments{
\item{x}{Object}
}
\value{
True if x is a grid_partition
}
\description{
is.grid_partition
}
