% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_plans.R
\name{lm_est}
\alias{lm_est}
\alias{is_lm_est}
\title{create an lm_est object}
\usage{
lm_est(lasso = FALSE, control_est = TRUE)

is_lm_est(x)
}
\arguments{
\item{lasso}{Use Lasso on the train sample to choose control variables}

\item{control_est}{Use controls when estimating effects on estimation sample}

\item{x}{an R object}
}
\value{
Estimation plan

Boolean
}
\description{
An estimator that can potentially control linearly for some features.

Tests whether object is an \code{lm_est}.
}
\section{Functions}{
\itemize{
\item \code{is_lm_est}: is lm_est
}}

