% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{num_cells.estimated_partition}
\alias{num_cells.estimated_partition}
\title{num_cells.estimated_partition}
\usage{
\method{num_cells}{estimated_partition}(obj)
}
\arguments{
\item{obj}{Estimated Partition}
}
\value{
Number of cells
}
\description{
num_cells.estimated_partition
}
