% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{num_cells.grid_partition}
\alias{num_cells.grid_partition}
\title{num_cells.grid_partition}
\usage{
\method{num_cells}{grid_partition}(obj)
}
\arguments{
\item{obj}{Object}
}
\value{
Number of cells
}
\description{
num_cells.grid_partition
}
