% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R, R/grid_partition.R, R/utils.R
\name{num_cells.estimated_partition}
\alias{num_cells.estimated_partition}
\alias{num_cells.grid_partition}
\alias{num_cells}
\title{Return number of cells for the object}
\usage{
\method{num_cells}{estimated_partition}(obj)

\method{num_cells}{grid_partition}(obj)

num_cells(obj)
}
\arguments{
\item{obj}{Object}
}
\value{
Number of cells in partition (at least 1)
}
\description{
Returns the number of cells for the an object
}
\section{Methods (by class)}{
\itemize{
\item \code{estimated_partition}: estimated_partition

\item \code{grid_partition}: grid_partition
}}

