% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Param_Est.grid_rf}
\alias{Param_Est.grid_rf}
\title{Param_Est.grid_rf}
\usage{
\method{Param_Est}{grid_rf}(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)
}
\arguments{
\item{obj}{Object}

\item{y}{y}

\item{d}{d}

\item{X}{X}

\item{sample}{Sample: "trtr", "trcv", "est"}

\item{ret_var}{Return Variance in the return list}
}
\value{
list(param_est=...)
}
\description{
Param_Est.grid_rf
}
