% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Param_Est.lm_X_est}
\alias{Param_Est.lm_X_est}
\title{Param_Est.lm_X_est}
\usage{
\method{Param_Est}{lm_X_est}(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)
}
\arguments{
\item{obj}{obj}

\item{y}{y}

\item{d}{d}

\item{X}{X}

\item{sample}{Sample: "trtr", "trcv", "est"}

\item{ret_var}{Return variance in return list}
}
\value{
list(param_est=...) or list(param_est=..., var_est=...)
}
\description{
Param_Est.lm_X_est
}
