% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{Param_Est}
\alias{Param_Est}
\title{Param_Est}
\usage{
Param_Est(obj, y, d = NULL, X, sample = "est", ret_var = FALSE)
}
\arguments{
\item{obj}{Object}

\item{y}{y A N-vector}

\item{d}{d A N-vector or Nxm matrix (so that they can be estimated jointly)}

\item{X}{X A NxK matrix or data.frame}

\item{sample}{Sample: "trtr", "trcv", "est"}

\item{ret_var}{Return Variance in the return list}
}
\value{
list(param_est=...)
}
\description{
Param_Est
}
