% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{partition_from_split_seq}
\alias{partition_from_split_seq}
\title{Generate a grid_partition from a sequence of splits}
\usage{
partition_from_split_seq(split_seq, X = NULL, X_range = NULL, varnames = NULL)
}
\arguments{
\item{split_seq}{list of splits}

\item{X}{Features. Must pass in this or \code{X_range}}

\item{X_range}{Domain for the features (saves time if already computed)}

\item{varnames}{variable names}
}
\value{
A \code{grid_partition}
}
\description{
Generate a grid_partition from a sequence of splits
}
