% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{predict.estimated_partition}
\alias{predict.estimated_partition}
\title{Generate predicted estimates per observations}
\usage{
\method{predict}{estimated_partition}(obj, new_X, new_d = NULL, type = "effect")
}
\arguments{
\item{obj}{estimated_partition object}

\item{new_X}{new X}

\item{new_d}{new d. Required for type="outcome"}

\item{type}{"effect" or "outcome" (currently not implemented)}
}
\value{
predicted treatment effect
}
\description{
Predicted unit-level treatment effect or outcome
}
