% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{predict.grid_partition}
\alias{predict.grid_partition}
\title{Get factor describing cell number fo each observation}
\usage{
\method{predict}{grid_partition}(obj, X, X_range = NULL)
}
\arguments{
\item{obj}{partition}

\item{X}{X data or list of X}

\item{X_range}{(Optional) overrides the partition$X_range}
}
\value{
Factor
}
\description{
Note that currently if X has values more extreme (e.g., for numeric or factor levels ) than was used to generate the partition
then we will return NA unless you provide and updated X_range.
}
