% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{predict_te.estimated_partition}
\alias{predict_te.estimated_partition}
\title{predict_te.estimated_partition}
\usage{
predict_te.estimated_partition(obj, new_X)
}
\arguments{
\item{obj}{estimated_partition object}

\item{new_X}{new X}
}
\value{
predicted treatment effect
}
\description{
Predicted unit-level treatment effect
}
