% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{print.estimated_partition}
\alias{print.estimated_partition}
\title{print.estimated_partition}
\usage{
\method{print}{estimated_partition}(
  x,
  do_str = TRUE,
  drop_unsplit = TRUE,
  digits = NULL,
  import_order = FALSE,
  ...
)
}
\arguments{
\item{x}{estimated_partition object}

\item{do_str}{If True, use a string like "(a, b]", otherwise have two separate columns with a and b}

\item{drop_unsplit}{If True, drop columns for variables overwhich the partition did not split}

\item{digits}{digits options}

\item{import_order}{should we use importance ordering or input ordering (default)}

\item{...}{Additional arguments. These won't be passed to print.data.frame}
}
\value{
string (and displayed)
}
\description{
print.estimated_partition
}
