% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{print.grid_partition}
\alias{print.grid_partition}
\title{Print grid_partition}
\usage{
\method{print}{grid_partition}(x, do_str = TRUE, drop_unsplit = TRUE, digits = NULL, ...)
}
\arguments{
\item{x}{partition object}

\item{do_str}{If True, use a string like "(a, b]", otherwise have two separate columns with a and b}

\item{drop_unsplit}{If True, drop columns for variables overwhich the partition did not split}

\item{digits}{digits Option}

\item{...}{Additional arguments. Passed to data.frame}
}
\value{
string (and displayed)
}
\description{
Prints a data.frame with options
}
