% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{print.partition_split}
\alias{print.partition_split}
\title{Print partition_split}
\usage{
\method{print}{partition_split}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Additional arguments. Unused.}
}
\value{
None
}
\description{
Prints information for a \code{partition_split}
}
