% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{quantile_breaks}
\alias{quantile_breaks}
\title{quantile_breaks}
\usage{
quantile_breaks(X, binary_k = c(), g = 20, type = 3)
}
\arguments{
\item{X}{Features}

\item{binary_k}{vector of dimensions that are binary}

\item{g}{# of quantiles}

\item{type}{Quantile type (see ?quantile and https://mathworld.wolfram.com/Quantile.html). 
Types1-3 are discrete and this is good for passing to unique() when there are clumps}
}
\value{
list of potential breaks
}
\description{
quantile_breaks
}
