% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{residualize}
\alias{residualize}
\title{Residualize sample}
\usage{
residualize(obj, y, X, d, sample)
}
\arguments{
\item{obj}{an EstimatorPlan object}

\item{y}{N-vector}

\item{X}{NxK matrix}

\item{d}{NULL, N-vector, or NxM matrix (if multiple treatments)}

\item{sample}{one of \code{'tr'} or \code{'est'}}
}
\value{
\code{list(y=, d=)}
}
\description{
Residualize y and d (only called when doing treatment effect estimation)
}
