% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalGrid.R
\docType{package}
\name{CausalGrid}
\alias{CausalGrid}
\title{CausalGrid: A package for subgroup effects}
\description{
Tools for finding heterogeneous treatment effects (and means) based on
partitioning the covariate/feature space via full cross-cuts and solved via
greedy search. A typical usage would be analyzing and experiment to find the
high-level subgroups (a coarse partition that is useful to humans) that
differ in their estimated treatment effects.
}
\details{
This package is inspired by, and uses ideas from, \code{Causal
Tree} but aims to have the
partition be more interpretable and have better accuracy. It is slower,
though for high-level partitions this is usually not an issue.

Subgroups are constructed as a grid over the features/covariates \code{X}.
For example, with 1 feature going from 0 to 1 it may split at values c1, c2,
resulting in segments \code{[0,c1], (c1, c2], (c2,1]}. A split at value
\code{c} means it splits <= and >. The segments may be of uneven sizes.
Splits along several features result in a grid by constructing the Cartesian
product of the feature-specific splits. Not all features will necessarily be
split or split the same number of times.

The main entry point is \code{\link{fit_estimate_partition}}.

Randomization: This package should be able to be run with no randomness. With
default/simple parameters the following places randomize but can be
overridden. \itemize{ 
\item  Generating train/est splits. Can be overridden
by providing \code{tr_split} 
\item  Generating trtr/trcv splits. Can be
overridden by providing \code{cv_folds} 
\item  Bumping samples. Can be
overridden by providing list of samples for \code{bump_samples} 
\item
Estimation plans: Provide ones ( \code{lm_est(lasso=TRUE,...)} and
\code{grid_rf}) use  \code{cv_folds}. User-made ones should too. 
}
}
