% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{change_complexity}
\alias{change_complexity}
\title{Change the complexity of a fit_estimate_partition}
\usage{
change_complexity(
  fit,
  y,
  X,
  d = NULL,
  partition_i,
  index_tr = fit$index_tr,
  split_seq = fit$split_seq,
  est_plan = fit$est_plan
)
}
\arguments{
\item{fit}{estimated_partition}

\item{y}{Nx1 matrix of outcome (label/target) data. With multiple core estimates see Details below.}

\item{X}{NxK matrix of features (covariates). With multiple core estimates see Details below.}

\item{d}{(Optional) NxP matrix (with colnames) of treatment data. If all equally important they 
should be normalized to have the same variance. With multiple core estimates see Details below.}

\item{partition_i}{partition_i - 1 is the last include in split_seq included in new partition}

\item{index_tr}{Split between train and estimate samples (default is to get from \code{fit})}

\item{split_seq}{sequential list of splits (default is to get from \code{fit})}

\item{est_plan}{\link{EstimatorPlan}.}
}
\value{
updated estimated_partition
}
\description{
Change the complexity level of the partition and re-estimate cell statistics. 
If you have a minimal estimated_partition then you need to pass in the other params.
}
\details{
Note: doesn't update the importance weights
}
\section{Multiple estimates}{

With multiple core estimates (M) there are 3 options (the first two have the same sample across treatment effects).\enumerate{
 \item DS.MULTI_SAMPLE: Multiple pairs of (Y_{m},W_{m}). y,X,d are then lists of length M. Each element then has the typical size
    The N_m may differ across m. The number of columns of X will be the same across m.
 \item DS.MULTI_D: Multiple treatments and a single outcome. d is then a NxM matrix.
 \item DS.MULTI_Y: A single treatment and multiple outcomes. y is then a NXM matrix.
}
}

