% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{change_complexity}
\alias{change_complexity}
\title{change_complexity}
\usage{
change_complexity(fit, y, X, d = NULL, partition_i)
}
\arguments{
\item{fit}{estimated_partition}

\item{y}{Nx1 matrix of outcome (label/target) data}

\item{X}{NxK matrix of features (covariates). Must be numerical (unordered categorical 
variables must be 1-hot encoded.)}

\item{d}{(Optional) NxP matrix (with colnames) or vector of treatment data. If all equally 
important they should be normalized to have the same variance.}

\item{partition_i}{partition_i - 1 is the last include in split_seq included in new partition}
}
\value{
updated estimated_partition
}
\description{
Doesn't update the importance weights
}
