% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{est_cell_stats}
\alias{est_cell_stats}
\title{Estimate parameters across the cells}
\usage{
est_cell_stats(
  y,
  X,
  d = NULL,
  partition = NULL,
  cell_factor = NULL,
  estimator_var = NULL,
  est_plan = NULL,
  alpha = 0.05
)
}
\arguments{
\item{y}{Nx1 matrix of outcome (label/target) data}

\item{X}{NxK matrix of features (covariates)}

\item{d}{(Optional) NxP matrix (with colnames) of treatment data. If all equally important they should 
be normalized to have the same variance.}

\item{partition}{(Optional, need this or cell_factor) partitioning returned from fit_estimate_partition}

\item{cell_factor}{(Optional, need this or partition)}

\item{estimator_var}{(Optional) a function with signature list(param_est, var_est) = function(y, d) 
(where if no d then can pass in null). If NULL then will choose between built-in 
mean-estimator and scalar_te_estimator}

\item{est_plan}{Estimation plan}

\item{alpha}{Alpha}
}
\value{
list
\item{cell_factor}{Factor with levels for each cell for X. Length N.}
\item{stats}{data.frame(cell_i, N_est, param_ests, var_ests, tstats, pval, ci_u, ci_l, p_fwer, p_fdr)}
}
\description{
Estimate the parameters (including standard errors) across the cells in the sample.
}
