% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{est_full_stats}
\alias{est_full_stats}
\title{est_full_stats}
\usage{
est_full_stats(
  y,
  d,
  X,
  est_plan,
  y_es = NULL,
  d_es = NULL,
  X_es = NULL,
  index_tr = NULL,
  alpha = 0.05
)
}
\arguments{
\item{y}{y}

\item{d}{d}

\item{X}{X}

\item{est_plan}{est_plan}

\item{y_es}{y_es}

\item{d_es}{d_es}

\item{X_es}{X_es}

\item{index_tr}{index_tr}

\item{alpha}{alpha}
}
\value{
Stats df
}
\description{
est_full_stats
}
