% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{estimated_partition}
\alias{estimated_partition}
\title{Minimum estimated_partition class}
\usage{
estimated_partition(partition, cell_stats, ...)
}
\arguments{
\item{partition}{partition}

\item{cell_stats}{cell_stats}

\item{...}{Additional arguments}
}
\value{
object of class estimated_partition
}
\description{
Minimum estimated_partition class
}
