% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{eval_mse_hat}
\alias{eval_mse_hat}
\title{Evaluate the MSE_hat objective function}
\usage{
eval_mse_hat(
  y_tr,
  X_tr,
  d_tr,
  y_te = NULL,
  X_te = NULL,
  d_te = NULL,
  N_est,
  partition = NULL,
  cell_factor_tr = NULL,
  cell_factor_te = NULL,
  est_plan = NULL,
  estimator = NULL,
  debug = FALSE,
  warn_on_error = FALSE,
  sample = "trtr",
  ...
)
}
\arguments{
\item{y_tr}{y_tr}

\item{X_tr}{X_tr}

\item{d_tr}{d_tr}

\item{y_te}{y_te}

\item{X_te}{X_te}

\item{d_te}{d_te}

\item{N_est}{Size of estimation sample. Unused for this objective function.}

\item{partition}{Grid partition. Pass in this or \code{cell_factor}.}

\item{cell_factor_tr}{Factor for cells for each observation. Pass in this or \code{partition}.}

\item{est_plan}{Estimation plan. If this and  \code{estimator} are null, then one is created as  \code{gen_simple_est_plan(has_d=!is_NULL_m(d))}}

\item{estimator}{If not passing in \code{est_plan}, can pass in this estimation routine and one is created using \code{\link{simple_est}}.}

\item{debug}{T/F whether we are in debug mode (and print out more info)}

\item{warn_on_error}{T/F for whether to display a warning when estimation fails (NA values returned)}

\item{sample}{Passed to \code{\link{est_params}}}
}
\value{
\code{c(val, N_cell_empty, N_cell_error)} val will be negative (and we're typically minimizing)
}
\description{
Evaluate the MSE_hat objective function over the cells for the sample
}
\keyword{internal}
