% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{fit_partition}
\alias{fit_partition}
\title{fit_partition}
\usage{
fit_partition(
  y,
  X,
  d = NULL,
  N_est = NA,
  X_aux = NULL,
  d_aux = NULL,
  max_splits = Inf,
  max_cells = Inf,
  min_size = 3,
  cv_folds = 2,
  verbosity = 0,
  pot_break_points = NULL,
  potential_lambdas = NULL,
  X_range = NULL,
  lambda.1se = FALSE,
  bucket_min_n = NA,
  bucket_min_d_var = FALSE,
  obj_fn,
  est_plan,
  partition_i = NA,
  pr_cl = NULL,
  valid_fn = NULL,
  bump_B = 0,
  bump_ratio = 1
)
}
\arguments{
\item{y}{Nx1 matrix of outcome (label/target) data}

\item{X}{NxK matrix of features (covariates)}

\item{d}{(Optional) NxP matrix (with colnames) of treatment data. If all equally important they 
should be normalized to have the same variance.}

\item{N_est}{N of samples in the Estimation dataset}

\item{X_aux}{aux X sample to compute statistics on (OOS data)}

\item{d_aux}{aux d sample to compute statistics on (OOS data)}

\item{max_splits}{Maximum number of splits even if splits continue to improve OOS fit}

\item{max_cells}{Maximum number of cells even if more splits continue to improve OOS fit}

\item{min_size}{Minimum cell size when building full grid, cv_tr will use (F-1)/F*min_size, cv_te doesn't use any.}

\item{cv_folds}{Number of Folds}

\item{verbosity}{If >0 prints out progress bar for each split}

\item{pot_break_points}{k-dim list of vectors giving potential split points}

\item{potential_lambdas}{potential lambdas to search through in CV}

\item{X_range}{list of min/max for each dimension}

\item{lambda.1se}{Use the 1se rule to pick the best lambda}

\item{bucket_min_n}{Minimum number of observations needed between different split checks}

\item{bucket_min_d_var}{Ensure positive variance of d for the observations between different split checks}

\item{obj_fn}{Whether to use the emse_hat or mse_hat}

\item{est_plan}{Estimation plan.}

\item{partition_i}{Default NA. Use this to avoid CV}

\item{pr_cl}{Default NULL. Parallel cluster (used for fit_partition_full)}

\item{valid_fn}{Function to quickly check if partition could be valid. User can override.}

\item{bump_B}{Number of bump bootstraps}

\item{bump_ratio}{For bootstraps the ratio of sample size to sample (between 0 and 1, default 1)}
}
\value{
list(partition, lambda)
}
\description{
CV Fit partition on some data, finds best lambda and then re-fits on full data.
}
