% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{get_desc_df.estimated_partition}
\alias{get_desc_df.estimated_partition}
\title{Get descriptive data.frame for an estimated_partition}
\usage{
get_desc_df.estimated_partition(
  obj,
  do_str = TRUE,
  drop_unsplit = TRUE,
  digits = NULL,
  import_order = FALSE
)
}
\arguments{
\item{obj}{estimated_partition object}

\item{do_str}{If True, use a string like "(a, b]", otherwise have two separate columns with a and b}

\item{drop_unsplit}{If True, drop columns for variables overwhich the partition did not split}

\item{digits}{digits Option (default is NULL)}

\item{import_order}{should we use importance ordering or input ordering (default)}
}
\value{
data.frame
}
\description{
Get statistics for each cell (feature boundary, and estimated cell stats)
}
