% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{get_factor_from_partition}
\alias{get_factor_from_partition}
\title{get_factor_from_partition}
\usage{
get_factor_from_partition(partition, X, X_range = NULL)
}
\arguments{
\item{partition}{partition}

\item{X}{X data or list of X}

\item{X_range}{(Optional) overrides the partition$X_range}
}
\value{
Factor
}
\description{
get_factor_from_partition
}
