% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{get_X_range}
\alias{get_X_range}
\title{Get X_range}
\usage{
get_X_range(X)
}
\arguments{
\item{X}{data}
}
\value{
list of length K with each element being the "range" along that dimension
}
\description{
Gets the "range" of each variable in X. For numeric variables this is (min, max).
For factors this means vector of levels.
}
\keyword{internal}
