% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{grid_rf}
\alias{grid_rf}
\title{grid_rf}
\usage{
grid_rf(num.trees = 500, num.threads = NULL, dof = 2, resid_est = TRUE)
}
\arguments{
\item{num.trees}{number of trees in the random forest}

\item{num.threads}{num.threads}

\item{dof}{degrees-of-freedom}

\item{resid_est}{Residualize the Estimation sample (using fit from training)}
}
\value{
grid_rf object
}
\description{
grid_rf
}
