% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{partition_split}
\alias{partition_split}
\title{Create partition_split}
\usage{
partition_split(k, X_k_cut)
}
\arguments{
\item{k}{dimension}

\item{X_k_cut}{cut value}
}
\value{
Partition Split
}
\description{
Describes a single partition split. Used with \code{\link{add_partition_split}}.
}
