% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot.estimated_partition}
\alias{plot.estimated_partition}
\title{Create 2D plots of parameter estimates}
\usage{
\method{plot}{estimated_partition}(grid_fit, X_names_2D = NULL)
}
\arguments{
\item{grid_fit}{grid_fit}

\item{X_names_2D}{X_names_2D}
}
\value{
ggplot2 object or list of such objects
}
\description{
Creates a 2D plot of parameter estimates or a series of such slices if partition is across >2 features.
}
