% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot_2D_partition.estimated_partition}
\alias{plot_2D_partition.estimated_partition}
\title{plot_2D_partition.estimated_partition}
\usage{
plot_2D_partition.estimated_partition(grid_fit, X_names_2D)
}
\arguments{
\item{grid_fit}{grid_fit}

\item{X_names_2D}{X_names_2D}
}
\value{
ggplot2 object
}
\description{
plot_2D_partition.estimated_partition
}
