% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_estimate.R
\name{print.estimated_partition}
\alias{print.estimated_partition}
\title{Print estimated_partition}
\usage{
\method{print}{estimated_partition}(
  x,
  do_str = TRUE,
  drop_unsplit = TRUE,
  digits = NULL,
  import_order = FALSE,
  ...
)
}
\arguments{
\item{x}{estimated_partition object}

\item{do_str}{If True, use a string like "(a, b]", otherwise have two separate columns with a and b}

\item{drop_unsplit}{If True, drop columns for variables overwhich the partition did not split}

\item{digits}{digits Option (default is NULL)}

\item{import_order}{Should we use importance ordering (most important on the left) or input ordering (default) for features.
Rows will be ordered so that the right-most will change most frequently.}

\item{...}{Additional arguments. These will be passed to print.data.frame}
}
\value{
string (and displayed)
}
\description{
Print a summary of the estimated partition. Uses \code{\link{get_desc_df.estimated_partition}}
}
