% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_plans.R
\name{simple_est}
\alias{simple_est}
\alias{is.simple_est}
\title{Create a simple_est object}
\usage{
simple_est(te_fn, te_var_fn, dof = 2)

is.simple_est(x)
}
\arguments{
\item{te_fn}{Bare Treatment Estimation function}

\item{te_var_fn}{Treatment estimatino function when variance needed}

\item{dof}{Degree-of-freedom correction (2 for intercept and treatment)}

\item{x}{an R object}
}
\value{
Estimation Plan

Boolean
}
\description{
A wrapper around a pair of estimator functions (w/ and w/o variance estimates)

Tests whether oject is an \code{simple_est}.
}
\section{Functions}{
\itemize{
\item \code{is.simple_est}: is simple_est
}}

